separating application logic from test units
