+++ lib/terminal/test/application/readme.md
@@ -4,6 +4,8 @@ The test runner utility and supporting libraries.
 ## Files
 <!-- Do not edit below this line.  Contents dynamically populated. -->
 
+* **[browser.ts](browser.ts)**       - The functions necessary to run browser test automation.
 * **[complete.ts](complete.ts)**     - Final messaging for a completed test type.
 * **[evaluation.ts](evaluation.ts)** - Evaluate a given test item and report appropriate failure messaging.
 * **[runner.ts](runner.ts)**         - A test runner that loops through test items in serial, executes those test items, and passes the result message to the evaluation library.
+* **[service.ts](service.ts)**       - A list of service test related utilities.
