+++ lib/terminal/test/samples/readme.md
@@ -5,5 +5,6 @@ The files that contain test samples and any sample type specific utilities.
 <!-- Do not edit below this line.  Contents dynamically populated. -->
 
 * **[browser.ts](browser.ts)**       - A list of tests that execute in the web browser.
-* **[service.ts](service.ts)**       - A list of service related tests.
+* **[service.ts](service.ts)**       - A list of service tests.
+* **[simulation.ts](simulation.ts)** - A list of command related tests for running shell simulations against the supported commands.
 * **[simulation.ts](simulation.ts)** - A list of command related tests for running shell simulations against the supported commands.
